deps.edn: Update versions, remove piggieback
I don't _think_ piggieback is needed for a pure clojure repl - I'm
assuming I was trying to get a cljs repl working, but 99% of the time
that's not what I'm doing with this alias. If I start using cljs more,
maybe I'll split out a second alias for a cljs repl.