nvim: Switch to eslint-ls
This removes the need for ALE doing eslint. I'm in the process of
removing `xo` from my projects, and don't _think_ I need prettier in
there anymore, so remove it for now.

There's still a bit of work to be done here - ideally, I'd be able to
:LspInstall eslint, and then it would automatically use the installed
language server binary, but currently I have to add the eslint-ls to my
path, which is janky. Also, the augroup that does the `eslint --fix`
seems to block the file from being written (or at least updating the
modified indicator? not sure).