nvim: Revert telescope, again
The matching algorithm is just ... not good?

If I'm in a repo and I search for "FunctionName(", once I type the "(",
all the matches go away?? And if there are extra spaces between two
words and you search for both words it doesn't work? Some questionable
matching, to be sure.