nvim: Remove vim-goimports in favor of lsp-based formatting
Resisted this for awhile, but seems reasonable at this point - _does_
suffer from the startup time any LSP server, but oh well. Largely based
on some of the suggestions in [this thread][1].

The formatting _does_ now run on any LSP-enabled buffer, which is
different from before. I had turned this off a very long time ago due to
constantly running into unformatted files and accidentally adding them
to my commits. I'll see how this run goes, and disable it if it's too
annoying.

[1]: https://github.com/neovim/nvim-lspconfig/issues/115