nvim: Set grepprg / grepformat for `rg`
I occasionally find it useful to use `:grep` to populate the quickfix
list with results, to have a more temporary buffer than fzf to see
results. I used to use vim-grepper, which I still _might_ bring back
that did this a bit cleaner, but for now, just set up grepprg /
grepformat for ripgrep.