nvim: Add packer.startup config
By default, the update buffer opens in a vertical split, which usually
doesn't fit on the screen quite right. This opens the update buffer in a
floating window instead. A bit weird due to the signature of the
`packer.startup` function, but luckily this is a good example for
elsewhere where the lua function accepts a mixed associative and
sequential table.