nvim: Remove vim-test (+ vim-dispatch)
I don't use vim-test enough to warrant keeping it around, and I've been
tempted to try out https://github.com/nvim-neotest/neotest - for now,
just remove it.

This also removes vim-dispatch, which I don't _think_ is used for
anything now (I believe it used to be used in fugitive, but it looks
like that behavior was removed[1]). Remove it for now.

[1]: https://github.com/tpope/vim-fugitive#faq