nvim: Add b0o/SchemaStore.nvim
Giving this a whirl, which needs jsonls & yamlls.

I'm fairly certain that jsonls is the [one pulled from vscode][1], but
despite saying that it supports code completion from schemas (and hover
working), it doesn't seem to expose that in its server_capabilities,
meaning that it isn't set up on the neovim side automatically. :shrug:.

[yamlls][2] doesn't seem all that well maintained, unfortunately, but
again, just including it for now and we'll see if it ends up being
useful.

Lastly, removed "handlers" from the setup of the servers, as that
variable no longer exists.

[1]: https://github.com/microsoft/vscode/tree/main/extensions/json-language-features/server
[2]: https://github.com/redhat-developer/yaml-language-server