fish: Set dummy repo_color variable
This is actually a workaround for situations without git repos (which
I'm in somewhat rarely now due to $HOME being a git repo for my
dotfiles). Without this, when not in a git repo, $repo_info variable is
empty, meaning that the `printf` only thinks it receives 2
arguments. This would place the `;` as the second argument, before the
\r, which would screw up the prompt.