nvim: Fix nfnl automatic compilation
I _think_ something was up with the filetype causing the plugin not to
load early enough to register its autocmds, or something?

The way that I fixed this was consistently checking
`:autocmd BufWritePost`. nfnl registers an autocmd to compile the fennel
files to lua, and if the autocmd is set up, everything is working.