fish: Replace `ls` with `eza`
I originally had backed out using `exa` in
e1781af28cbe8d07336664bd9f22301f7ebf6b84, but now that it's back to
being maintained in a fork, give it another shot. Slightly nicer output,
and support for a tree view built-in (which I occasionally install
`tree` for) means that this is at least something to try.