kitty: Add open last command output in nvim shortcut
Every once in awhile I'll run a command, then realize the output is too
much, so I'll re-run it with `| vim -R -` at the end to get the output
in vim.

I figured kitty would have a better solution to this, and it does:
https://sw.kovidgoyal.net/kitty/conf/#shortcut-kitty.Browse-output-of-the-last-shell-command-in-pager

Use a similar key combination to the other kitty shortcuts, for
consistency.