nvim: Remove vim-test mappings
The issue with these mappings is that <C-t> is used to pop back out of
the "tag" stack (when using <C-]> / gd). Having these mappings cause vim
to wait to determine if the second key will be pressed.

Have been wondering about this for awhile; I've gotten into the habit of
using <C-o> for popping the stack, but that's not _really_ what it does
and it can be annoying when trying to trace back out.

May end up remapping these to something else, but I mostly just use
`:TestNearest`, and that's only on occasion.