nvim: Lean into conform formatting
This removes the LSP-specific formatting / autocmd I was setting up in
favor of conform.nvim. This means that conform will run its specified
formatter, and if one isn't provided, fallback to the LSP formatting.

This came up in the situation where both `tsserver` (LSP) and `prettier`
(conform) were fighting over formatting a given buffer - by doing it
this way, the filetype in conform will be respected, and then fallback
to any configured LSP formatter.

As a side-effect of this, enable `goimports` for go files, since it was
special-cased in the format function to start with. (Ideally, would
figure out how to avoid this, but not a big deal.)