nvim: Switch back to LSP document highlighting
mini-cursorword is _okay_, but it's not nearly as precise as using
LSP. (It sucks that `gopls` has such a long startup time (even for the
client using a shared daemon - it seems to be bound by `go list` being
slow AF).) modus also has a hideous color with cursorword by default
that I didn't investigate more. This brings back the various document
highlight augroup, autocmd, and updatetime settings (using 300 instead
of 100 as it seems more reasonable for updates).