git: Fix mergetool invocation
But tbh, this still isn't quite right as the configured `mergetool`,
because it'll open all the files needing merge in the quickfix list, but
as they're resolved, it'll then open each file in turn. Seems like I
ought to either do this or _separately_ open `nvim` and use `:Git
mergetool`.